requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg==3.2.6
PyJWT==2.9.0
sqlparse==0.5.3
Pillow==10.2.0
django-cors-headers==4.3.1
firebase-admin==6.4.0
